Stops building unused NGINX image
TL;DR
-----

Removes workflow steps related to NGINX image

Details
-------

Cleans up the workflow so it no longer does anything with the NGINX
image that hasn't been used for a long time. The chart stopped using the
imagea while ago and I updated the upstream to account for that last
week.